+++ xplat/src/main/java/lgbt/greenhouse/enchantmentdisabletag/duck/Duck_PotentialEnchantmentDisabledStack.java
@@ -1,0 +1,6 @@
+package lgbt.greenhouse.enchantmentdisabletag.duck;
+
+public interface Duck_PotentialEnchantmentDisabledStack {
+    boolean enchantmentdisabletag$wasDisabled();
+    void enchantmentdisabletag$setWasDisabled();
+}