+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 ## Changes
 - Simplified the structure of the entire mod.
+- The mod will now remove obsolete disabled items within the creative mode tab.
 - The mod will only remove disabled enchantments when applied directly to an ItemStack. This allows innate enchantments to work out of the box.
+
+## Bugfixes
+- Fixed creative mode items with disabled enchantments causing crashes when two of the same are present. (Related to removing obsolete items).